+++ src/api/volume.go
@@ -178,9 +178,6 @@ func (s *VolumeService) Attach(ctx context.Context, volume *csi.Volume, server *
 		if hcloud.IsError(err, hcloud.ErrorCode("limit_exceeded_error")) {
 			return volumes.ErrAttachLimitReached
 		}
-		if hcloud.IsError(err, hcloud.ErrorCodeLocked) {
-			return volumes.ErrLockedServer
-		}
 		return err
 	}
 
@@ -240,9 +237,6 @@ func (s *VolumeService) Detach(ctx context.Context, volume *csi.Volume, server *
 				"server-id", server.ID,
 				"err", err,
 			)
-			if hcloud.IsError(err, hcloud.ErrorCodeLocked) {
-				return volumes.ErrLockedServer
-			}
 			return err
 		}
 		if hcloudServer == nil {